dock strings
